Updates to documentation and some scenario paths in examples and integration tests based on 2.0 reorganization of files
