+++ docs/scenarios/structural.rst
@@ -2,7 +2,7 @@
 Structural Scenario
 %%%%%%%%%%%%%%%%%%%
 
-The :class:`ScenarioStructural<mphys.scenario_structural.ScenarioStructural>` is for scenarios that only need the structural solver.
+The :class:`ScenarioStructural<mphys.scenarios.structural.ScenarioStructural>` is for scenarios that only need the structural solver.
 An examples use case for this scenario would be a stress constraint from a static load generated by the structural solver subsystem(s).
 
 Default Solvers
@@ -12,7 +12,7 @@ The default solvers are NonlinearRunOnce and LinearRunOnce that execute the pre 
 Options
 =======
 .. embed-options::
-  mphys.scenario_structural
+  mphys.scenarios.structural
   ScenarioStructural
   options
