+++ examples/aero_only/mach_tutorial_wing/adflow/mphys_aero_opt_2pt.py
@@ -3,7 +3,7 @@ import numpy as np
 
 import openmdao.api as om
 from mphys.multipoint import Multipoint
-from mphys.scenario_aerodynamic import ScenarioAerodynamic
+from mphys.scenarios.aerodynamic import ScenarioAerodynamic
 from adflow.mphys import ADflowBuilder
 from baseclasses import AeroProblem
 from pygeo.mphys import OM_DVGEOCOMP
