+++ examples/aero_only/mach_tutorial_wing/vlm/run_vlm_2scenarios_parallel.py
@@ -3,7 +3,7 @@ import openmdao.api as om
 
 from mphys import MultipointParallel
 from vlm_solver.mphys_vlm import VlmBuilder
-from mphys.scenario_aerodynamic import ScenarioAerodynamic
+from mphys.scenarios.aerodynamic import ScenarioAerodynamic
 
 class ParallelCruises(MultipointParallel):
     def setup(self):
