+++ examples/aerostructural/supersonic_panel/run_parallel.py
@@ -3,7 +3,7 @@ from mpi4py import MPI
 import openmdao.api as om
 
 from mphys import Multipoint, MultipointParallel
-from mphys.scenario_aerostructural import ScenarioAeroStructural
+from mphys.scenarios.aerostructural import ScenarioAeroStructural
 
 from structures_mphys import StructBuilder
 from aerodynamics_mphys import AeroBuilder
