+++ examples/struct_only/beam_shape_opt/shape_opt.py
@@ -30,7 +30,7 @@ import openmdao.api as om
 from pygeo.mphys import OM_DVGEOCOMP
 
 from mphys import Multipoint
-from mphys.scenario_structural import ScenarioStructural
+from mphys.scenarios.structural import ScenarioStructural
 from tacs.mphys import TacsBuilder
 from tacs import elements, constitutive, functions
