+++ tests/integration_tests/test_dafoam_derivs.py
@@ -19,7 +19,7 @@ import openmdao.api as om
 from openmdao.utils.assert_utils import assert_near_equal
 
 from mphys.multipoint import Multipoint
-from mphys.scenario_aerodynamic import ScenarioAerodynamic
+from mphys.scenarios.aerodynamic import ScenarioAerodynamic
 
 baseDir = os.path.dirname(os.path.abspath(__file__))
 dataDir = baseDir + "/../input_files/dafoam_aero/"
