+++ tests/regression_tests/test_dafoam_aero.py
@@ -22,7 +22,7 @@ import openmdao.api as om
 from openmdao.utils.assert_utils import assert_near_equal
 
 from mphys.multipoint import Multipoint
-from mphys.scenario_aerodynamic import ScenarioAerodynamic
+from mphys.scenarios.aerodynamic import ScenarioAerodynamic
 
 baseDir = os.path.dirname(os.path.abspath(__file__))
 dataDir = baseDir + "/../input_files/dafoam_aero/"
